tests/submission/PhilipTaylor/canvas/2d.line.defaults.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.line.defaults</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.defaults</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.lineWidth.default">2d.lineWidth.default</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.lineCap.default">2d.lineCap.default</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.lineJoin.default">2d.lineJoin.default</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.miterLimit.default">2d.miterLimit.default</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

_assertSame(ctx.lineWidth, 1, "ctx.lineWidth", "1");
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");
_assertSame(ctx.lineJoin, 'miter', "ctx.lineJoin", "'miter'");
_assertSame(ctx.miterLimit, 10, "ctx.miterLimit", "10");


});
</script>

<div id="log"></div>
</body></html>
